added some javadoc and moved timeseriesapptern methods from EntityNamingStrategy to FileNamingStrategy
