+++ src/main/java/edu/ie3/datamodel/io/naming/DefaultDirectoryHierarchy.java
@@ -44,7 +44,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /** Default directory hierarchy for input models */
-public class DefaultDirectoryHierarchy implements DirectoryHierarchy {
+public class DefaultDirectoryHierarchy implements FileHierarchy {
   private static final Logger logger = LoggerFactory.getLogger(DefaultDirectoryHierarchy.class);
 
   /** Use the unix file separator here. */
